Revert "move metabase-lib tests to create-metabase-query unit tests"
This reverts commit fde67d7f63f3690f98123da5679f50f6bb728691.
